--- a/Makefile
+++ b/Makefile
@@ -1,2 +1,0 @@
-run: 
-	go run local/*.go ${file}