--- a/go.mod
+++ b/go.mod
@@ -1,3 +1,0 @@
-module github.com/jorgenhanssen/a-machine
-
-go 1.16