--- a/programs/call
+++ b/programs/call
@@ -4,12 +4,12 @@ set '3' 0
 set '5' 1
 
 // We want to square the value in register 0
-set 0 100
-call SQUARE 
+set '0' 100
+call SQUARE
 
 // We want to square the value in register 1
-set 1 100
-call SQUARE 
+set '1' 100
+call SQUARE
 
 // Print the results
 out 0
@@ -22,4 +22,3 @@ term
 SQUARE:
 mul &100 &100 &100
 ret
-