--- a/src/main.rs
+++ b/src/main.rs
@@ -21,12 +21,6 @@ fn main() {
     let compile_time = compile_start_time.elapsed();
     println!("Compiled in {}μs", compile_time.as_micros());
 
-    for instruction in &instructions {
-        println!("{:?}", &instruction);
-    }
-
     let mut runner = runner::Runner::new(instructions);
     runner.run();
-
-    println!("Done");
 }