## Motivation


Let's face it, Rust is the new black. It's the fabric everything must be cut from if you want to be part of the cool code club. Now, one might be wondering, "Why on earth would someone re-write their own programming language's runtime in Rust instead of Go?" Well, I embarked on this Herculean task not just because Rust's safety scissors promise not to run with my code and trip over memory leaks, but because in the grand tradition of software development, I too wanted to blindly follow the herd. After all, if everyone jumps off the bridge, it's bound to get awfully lonely standing up there by yourself.

In the same way that fashion dictates we wear shoes on our hands just because it's Tuesday, the programming world has decided that Rust is the new speed daemon, capable of outrunning anything with its hands tied behind its back. So, I thought to myself, "Why not?" Why not rewrite everything in Rust, because apparently, it’s not just a programming language; it’s a lifestyle choice. A choice that says, "Yes, I too enjoy the finer things in life, like spending countless nights wrestling with the borrow checker instead of sleeping."

But let's not kid ourselves. The real reason to go through the painstaking process of rewriting the runtime in Rust is the same reason one would buy a sports car during a mid-life crisis. It's not about the speed, or the sleek syntax, or even the safety. It's about looking in the mirror and saying, "I am still relevant, dammit." It’s about embracing the existential programmer angst that comes with knowing that no matter how fast your code runs, someone, somewhere, is rewriting it in Rust for an extra 0.001% performance boost.

I rewrote my programming language's runtime in Rust not because I needed to, but because I could. Because in a world where performance is king, and memory safety is the queen, Rust is the royal flush in the high stakes poker game of programming languages. And if we're all going to be flushed away in the relentless pursuit of performance, we might as well do it with style and a compiler that cares deeply about our feelings. So, here’s to Rust, the reason we all get to pretend we’re avant-garde software artisans, crafting exquisitely efficient binaries that no one cares about. Cheers!

## Benchmark

The prime benchmark `programs/bench` is used as a benchmark program, where it calculates wether 50728129 is a prime number. This benchmark does not indicate all functionalities of the language, but serves as a general performance of the tick time.

There are some unfair advantages for the Rust runtime, as it does not generate a log throughout execution.

### Old runtime (Golang)
Compiled with `go build -ldflags="-s -w" local/*.go`

```
time ./compiler programs/bench
Compiled in 60.17µs
24432
./compiler programs/bench  26.59s user 3.86s system 122% cpu 24.845 total
```

### New runtime (Rust)
Compiled with `cargo build --release`

```
time target/release/src programs/bench
Compiled in 90μs
258
target/release/src programs/bench  0.26s user 0.00s system 99% cpu 0.262 total
```

### Benchmark improvement

It appears that the relative compilation time is quite higher for the Rust runtime, however, being in microseconds, the actual difference is negligible. Most importantly, the execution time for the prime-finder benchmark was improved by **9470%** using the Rust runtime, reducing the execution time from **24s to 258ms**.